refactor(workspace): separate persistence from sync extensions
Move persistence extensions (IndexedDB, SQLite) out of sync/ into their
own persistence/ directory. Rename files after their technology instead
of their platform (web.ts → indexeddb.ts, desktop.ts → sqlite.ts).

Before: extensions/sync/ contained 2/3 persistence files
After:  extensions/persistence/ for storage, extensions/sync/ for sync

Subpath exports updated:
  extensions/persistence/indexeddb (was extensions/sync/web)
  extensions/persistence/sqlite    (was extensions/sync/desktop)
  extensions/sync/websocket        (was extensions/sync)
  extensions/sync/broadcast-channel (unchanged)
