+++ apps/skills/src/lib/client.ts
@@ -1,4 +1,4 @@
 import { createSkillsWorkspace } from '@epicenter/skills';
-import { indexeddbPersistence } from '@epicenter/workspace/extensions/sync/web';
+import { indexeddbPersistence } from '@epicenter/workspace/extensions/persistence/indexeddb';
 
 export const workspace = createSkillsWorkspace().withExtension('persistence', indexeddbPersistence);
