+++ apps/whispering/src/lib/client.ts
@@ -4,7 +4,7 @@
  * Future sync extensions will add remote replication.
  */
 
-import { indexeddbPersistence } from '@epicenter/workspace/extensions/sync/web';
+import { indexeddbPersistence } from '@epicenter/workspace/extensions/persistence/indexeddb';
 import { createWhispering } from './workspace/workspace';
 
 export const workspace = createWhispering().withExtension('persistence', indexeddbPersistence);
