+++ docs/guides/consuming-epicenter-api.md
@@ -99,9 +99,9 @@ These fields only appear in `getSession()` responses, not in `signIn` or `signUp
 
 ```typescript
 import { createWorkspace } from '@epicenter/workspace';
-import { createSyncExtension } from '@epicenter/workspace/extensions/sync';
+import { createSyncExtension } from '@epicenter/workspace/extensions/sync/websocket';
 import { broadcastChannelSync } from '@epicenter/workspace/extensions/sync/broadcast-channel';
-import { indexeddbPersistence } from '@epicenter/workspace/extensions/sync/web';
+import { indexeddbPersistence } from '@epicenter/workspace/extensions/persistence/indexeddb';
 import { bytesToBase64 } from '@epicenter/workspace/shared/crypto';
 
 const client = createWorkspace(definition)
