+++ packages/cli/test/e2e-honeycrisp.test.ts
@@ -14,7 +14,7 @@ import { afterAll, beforeAll, describe, expect, test } from 'bun:test';
 import { rm } from 'node:fs/promises';
 import { join } from 'node:path';
 import { dateTimeStringNow } from '@epicenter/workspace';
-import { filesystemPersistence } from '@epicenter/workspace/extensions/sync/desktop';
+import { filesystemPersistence } from '@epicenter/workspace/extensions/persistence/sqlite';
 import { honeycrisp, createHoneycrisp } from '@epicenter/honeycrisp/workspace';
 import { loadConfig } from '../src/load-config';
