+++ packages/workspace/package.json
@@ -9,9 +9,9 @@
 			"node": "./src/extensions/index.ts",
 			"default": "./src/extensions/index.ts"
 		},
-		"./extensions/sync": "./src/extensions/sync.ts",
-		"./extensions/sync/web": "./src/extensions/sync/web.ts",
-		"./extensions/sync/desktop": "./src/extensions/sync/desktop.ts",
+		"./extensions/persistence/indexeddb": "./src/extensions/persistence/indexeddb.ts",
+		"./extensions/persistence/sqlite": "./src/extensions/persistence/sqlite.ts",
+		"./extensions/sync/websocket": "./src/extensions/sync/websocket.ts",
 		"./extensions/sync/broadcast-channel": "./src/extensions/sync/broadcast-channel.ts",
 		"./ingest": "./src/ingest/index.ts",
 		"./ingest/reddit": "./src/ingest/reddit/index.ts",
